Refactoring code and resolving review
